# Enable multithreading in hashrows_col!