--- a/benchmarks/grouping_performance.jl
+++ b/benchmarks/grouping_performance.jl
@@ -1,5 +1,4 @@
 using DataFrames
-using CategoricalArrays
 using PooledArrays
 using BenchmarkTools
 using Random
@@ -8,29 +7,50 @@ Random.seed!(1)
 
 grouping_benchmarks = BenchmarkGroup()
 
-# `refpool`/`refarray` optimized grouping method
-refpool_benchmarks = grouping_benchmarks["refpool"] = BenchmarkGroup()
-
-for k in (10, 10_000), n in (100, 100_000, 10_000_000)
-    for x in (PooledArray(rand(1:k, n)),
-              CategoricalArray(rand(1:n, 10_000_000)),
-              PooledArray(rand([missing; 1:n], 10_000_000)),
-              CategoricalArray(rand([missing; 1:n], 10_000_000)))
+for n in (1000, 10_000_000),
+    k in (10, 10_000),
+    x in (rand(1:k, n),
+          rand([1:k; missing], n),
+          rand(string.(1:k), n),
+          rand([string.(1:k); missing], n),
+          PooledArray(rand(1:k, n)),
+          PooledArray(rand([missing; 1:k], n)))
         df = DataFrame(x=x)
 
-        refpool_benchmarks[k, n, nameof(typeof(x)), "skipmissing=false"] =
+        grouping_benchmarks[n, k, typeof(x), "skipmissing=false"] =
             @benchmarkable groupby($df, :x)
 
         # Skipping missing values
-        refpool_benchmarks[k, n, nameof(typeof(x)), "skipmissing=true"] =
+        grouping_benchmarks[n, k, typeof(x), "skipmissing=true"] =
             @benchmarkable groupby($df, :x, skipmissing=true)
 
-        # Empty group which requires adjusting group indices
-        replace!(df.x, 5 => 6)
-        refpool_benchmarks[k, n, nameof(typeof(x)), "empty group"] =
-            @benchmarkable groupby($df, :x)
-    end
+        for l in (10, 10_000),
+            y in (rand(1:l, n),
+                  rand([1:l; missing], n),
+                  rand(string.(1:l), n),
+                  rand([string.(1:l); missing], n),
+                  PooledArray(rand(1:l, n)),
+                  PooledArray(rand([missing; 1:l], n)))
+            df.y = y
+
+            grouping_benchmarks[n, (k, l), (typeof(x), typeof(y)),
+                                "skipmissing=false"] =
+                @benchmarkable groupby($df, [:x, :y])
+
+            # Skipping missing values
+            grouping_benchmarks[n, (k, l), (typeof(x), typeof(y)),
+                                "skipmissing=true"] =
+                @benchmarkable groupby($df, [:x, :y], skipmissing=true)
+        end
+
+        if df.x isa PooledArray
+            # Empty group which requires adjusting group indices
+            replace!(df.x, levels(df.x)[1] => levels(df.x)[2])
+            grouping_benchmarks[n, k, typeof(x), "empty group"] =
+                @benchmarkable groupby($df, :x)
+        end
 end
+
 
 # If a cache of tuned parameters already exists, use it, otherwise, tune and cache
 # the benchmark parameters. Reusing cached parameters is faster and more reliable