--- a/docs/src/lib/internals.md
+++ b/docs/src/lib/internals.md
@@ -15,5 +15,6 @@ gennames
 getmaxwidths
 ourshow
 ourstrwidth
+@spawn_for_chunks
 tforeach
 ```