This is somewhat faster for large number of rows (by about 20% for 10M rows and 10 threads), but a bit slower when the number of rows is lower than the threshold (so that a single thread is used). Not sure whether something can be done about it, nor whether the improvement in some cases is worth the loss in others.

```julia
julia> using Revise, DataFrames, BenchmarkTools

julia> for n in (1000, 10_000_000)
           df = DataFrame(x=rand(string.(1000000:1000010), n));
           @btime groupby($df, :x);

          # Benchmark directly the changes
           h = similar(df.x, UInt);
           n = similar(df.x, Bool);
           @btime DataFrames.hashrows_col!($h, $n, $df.x, nothing, true);
       end

# main
  21.105 μs (29 allocations: 34.05 KiB)
  12.211 μs (0 allocations: 0 bytes)
  335.166 ms (32 allocations: 280.59 MiB)
  122.653 ms (0 allocations: 0 bytes)

# PR
  26.041 μs (29 allocations: 34.05 KiB)
  16.291 μs (0 allocations: 0 bytes)
  258.437 ms (107 allocations: 280.60 MiB)
  18.463 ms (74 allocations: 5.72 KiB)
```