@bkamins 2021-03-28
I was testing the same when implementing hashing in `OnCols` for joins and had similar conclusions. We probably can have it, but the threshold for turning "this on" should probably be large.