@nalimilan 2021-03-28
> I was testing the same when implementing hashing in `OnCols` for joins and had similar conclusions. We probably can have it, but the threshold for turning "this on" should probably be large.

The point is, the regression happens even under the threshold. It looks like we pay the cost even when multithreading isn't used.