@bkamins 2021-03-28
My guess is that we would have to split it into several methods. Maybe (I have not checked, but seen it in other cases), when you have very long methods that branch compiles is not always doing the right thing. If you will not find the root cause of this I can look at the issue later this week.