@nalimilan 2021-04-12
I've switched to using a macro (inspired by `@threads` but using `@spawn`) instead of a function, and the overhead is negligible now. Actually the syntax is even nicer, it's almost identical to the non-threaded version. I guess it could replace `tforeach`.

Updated benchmark (with 4 threads):
main:
```julia
  16.520 μs (29 allocations: 34.05 KiB)
  9.406 μs (0 allocations: 0 bytes)
  207.426 ms (32 allocations: 280.59 MiB)
  100.835 ms (0 allocations: 0 bytes)
```

PR:
```julia
  17.076 μs (29 allocations: 34.05 KiB)
  10.564 μs (0 allocations: 0 bytes)
  139.823 ms (93 allocations: 280.60 MiB)
  33.247 ms (61 allocations: 5.16 KiB)
```