@nalimilan 2021-04-12
I confirm that benchmarks give similar results with the fix.