@bkamins 2021-04-12
Looks good now. I leave up to you if you want to investigate https://github.com/JuliaData/DataFrames.jl/pull/2685/files#r611769396 still or merge it as is.