@nalimilan 2021-04-12
> Looks good now. I leave up to you if you want to investigate https://github.com/JuliaData/DataFrames.jl/pull/2685/files#r611769396 still or merge it as is.

Given that the current approach allows running both steps in parallel I'd say it's OK (and I had benchmarked it when it was introduced IIRC).